backend/requirements.txt
Django >= 4.2
djangorestframework >= 3.14
djangorestframework-simplejwt >= 5.5
django-cors-headers >= 4.3
psycopg2 >= 2.9
Pillow >= 12.1
python-magic >= 0.4.27
python-dotenv >= 1.0
